requirements.txt
django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary==2.9.9
django-redis==5.4.0
celery==5.4.0
redis==5.0.7
uvicorn==0.30.1
gunicorn==22.0.0
python-dotenv==1.0.1
django-environ==0.11.2
django-cors-headers==4.4.0
django-prometheus==2.3.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-requests==0.48b0
minio==7.2.7
httpx==0.27.0
PyJWT==2.8.0
cryptography==43.0.1

# Dev/test
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
schemathesis==3.35.0
black==24.8.0
ruff==0.5.7
mypy==1.11.1
types-redis==4.6.0.20240726
types-requests==2.32.0.20240712
pre-commit==3.8.0
